project/backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
edge-tts>=6.1.9
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.2.1
scipy>=1.11.0
